Client vanilla: add log logic as in the react version

+++ client-vanilla/src/main.ts
@@ -1,27 +1,57 @@
 import "./style.scss";
+import {Credentials} from "./credentials";
 import {MainPage} from "./main-page";
 import {DialogAddEdit} from "./dialog-add-edit";
+import {DialogLogin} from "./dialog-login";
 import {trpc} from "./trpc";
 
-
+export const auth = new Credentials();
 export const mainPage = new MainPage();
+export const dialogLogin = new DialogLogin();
 export const dialogAddEdit = new DialogAddEdit();
 
 
 // @1: init
 mainPage.setButtonsEnabled(false);
-const {token} = await trpc.login.mutate({username: "client-vanilla", password: "password"});
-window.addEventListener("beforeunload", () => {trpc.logout.mutate({token})});
+dialogLogin.setOnOkClicked(async (username: string, password: string) => {
+	try {
+		const {token} = await trpc.login.mutate({username, password});
+		auth.login(username, password, token);
+		dialogLogin.setVisible(false);
+
+		mainPage.clearAllEntries();
+		const entries = await trpc.getAllEntries.query();
+		for (const e of entries) {mainPage.appendEntry(e);}
+		mainPage.setButtonsEnabled(true);
+
+		// subscription that needs the token to be sent when subscribing
+		const onLoggedOutUnsubscribable = trpc.onLoggedOut.subscribe({token: auth.token()}, {
+			onData: () => {
+				auth.logout();
+				mainPage.setButtonsEnabled(false);
+				dialogAddEdit.setVisible(false);
+				dialogLogin.setVisible(true);
+				onLoggedOutUnsubscribable.unsubscribe();
+			}
+		});
+	}
+	catch (error) {
+		if (typeof error === "object" && error !== null && "message" in error) {alert(`Authentication failed:\n${error.message}`);}
+		else {alert(error);}
+	}
+});
+
+dialogAddEdit.setVisible(false);
+dialogLogin.setVisible(true);
+
+
+window.addEventListener("beforeunload", () => {trpc.logout.mutate({token: auth.token()})});
 // remark: the line above does not work
 // - on Chrome: sometimes, but not always (didn't get why)
 // - on Firefox: work if the tab being closed is not the last one in this Firefox window
 
-const entries = await trpc.getAllEntries.query();
-for (const e of entries) {mainPage.appendEntry(e);}
-mainPage.setButtonsEnabled(true);
 
-
-// @1: add / edit / remove buttons callbacks
+// @1: buttons callbacks (add, edit, remove)
 type Mode = "closed" | "open-add" | "open-edit";
 let mode: Mode = "closed";
 
@@ -38,7 +68,7 @@ async function handleEditClicked() {
 
 	try {
 		// lock the entry to update
-		await trpc.lock.mutate({entryId: selectedEntry.id, token});
+		await trpc.lock.mutate({entryId: selectedEntry.id, token: auth.token()});
 		dialogAddEdit.setVisible(true);
 
 		// fill and display the dialog
@@ -56,10 +86,10 @@ async function handleDeleteClicked() {
 
 	try {
 		// lock the entry to delete
-		await trpc.lock.mutate({entryId: selectedEntry.id, token});
+		await trpc.lock.mutate({entryId: selectedEntry.id, token: auth.token()});
 
 		// send delete request
-		await trpc.deleteEntry.mutate({entryId: selectedEntry.id, token});
+		await trpc.deleteEntry.mutate({entryId: selectedEntry.id, token: auth.token()});
 	}
 	catch (error) {
 		alert(error);
@@ -104,10 +134,10 @@ async function handleAddOk() {
 async function handleEditOk() {
 	try {
 		const entryFromDialog = dialogAddEdit.getData();
-		await trpc.updateEntry.mutate({...entryFromDialog, token});
+		await trpc.updateEntry.mutate({...entryFromDialog, token: auth.token()});
 	
 		// unlock the updated entry
-		await trpc.unlock.mutate({entryId: entryFromDialog.id, token});
+		await trpc.unlock.mutate({entryId: entryFromDialog.id, token: auth.token()});
 	}
 	catch (error) {
 		alert(error);
@@ -122,7 +152,7 @@ async function handleAddEditCancel() {
 		if (mode === "open-edit") {
 			// unlock the entry being edited
 			const entryFromDialog = dialogAddEdit.getData();
-			await trpc.unlock.mutate({entryId: entryFromDialog.id, token});
+			await trpc.unlock.mutate({entryId: entryFromDialog.id, token: auth.token()});
 		}
 	}
 	catch (error) {
@@ -138,28 +168,24 @@ dialogAddEdit.setOnOkClicked(handleAddEditOk);
 dialogAddEdit.setOnCancelClicked(handleAddEditCancel);
 
 
-// @1: notifications
-trpc.onLoggedOut.subscribe({token}, {
-	onData: () => {
-		console.log("Token expired!");
-		alert("Token expired!");
-	}
-});
-
+// @1: notifications that don't need the token to be sent when subscribing
 trpc.onEntryInserted.subscribe(undefined, {
 	onData: (newEntry) => {
+		if (!auth.isLoggedIn()) {return;}
 		mainPage.appendEntry(newEntry);
 	}
 });
 
 trpc.onEntryUpdated.subscribe(undefined, {
 	onData: (updatedEntry) => {
+		if (!auth.isLoggedIn()) {return;}
 		mainPage.updateEntry(updatedEntry);
 	}
 });
 
 trpc.onEntryDeleted.subscribe(undefined, {
 	onData: ({deletedEntryId}) => {
+		if (!auth.isLoggedIn()) {return;}
 		mainPage.deleteEntry(deletedEntryId);
 	}
 });
